Move tshd test helpers to a better location
