+++ web/packages/teleterm/src/services/tshd/testHelpers.ts
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 
-import { ResourceSearchResult } from './searchResult';
-
-import type * as tsh from 'teleterm/services/tshd/types';
+import type * as tsh from './types';
 
 export const makeServer = (props: Partial<tsh.Server>): tsh.Server => ({
   uri: '/clusters/teleport-local/servers/178ef081-259b-4aa5-a018-449b5ea7e694',
@@ -49,15 +47,3 @@ export const makeKube = (props: Partial<tsh.Kube>): tsh.Kube => ({
 
 export const makeLabelsList = (labels: Record<string, string>): tsh.Label[] =>
   Object.entries(labels).map(([name, value]) => ({ name, value }));
-
-export const makeResourceResult = (
-  props: Partial<ResourceSearchResult> & {
-    kind: ResourceSearchResult['kind'];
-    resource: ResourceSearchResult['resource'];
-  }
-): ResourceSearchResult => ({
-  score: 0,
-  labelMatches: [],
-  resourceMatches: [],
-  ...props,
-});
